requirements.txt
yfinance
pandas
requests
scikit-learn
python-dotenv
streamlit
plotly
gspread
google-auth
